seperate game logic in multiple files
