+++ app/composables/useGame.ts
@@ -1,9 +1,3 @@
-import {
-  activatePlayerCard,
-  activatePropertyCard,
-  createGame,
-  outOfJail,
-} from "~/utils/game";
 import { io } from "socket.io-client";
 
 let gameRef: Ref<Game> | undefined;
